Added changes to files and code
