+++ client/src/scenes/form/LoginForm.js
@@ -5,8 +5,8 @@ import { Box, Button, TextField, Typography } from "@mui/material";
 import { Formik } from "formik";
 import * as yup from "yup";
 import useMediaQuery from "@mui/material/useMediaQuery";
-import Divider from "@mui/material/Divider";
-import CreateAccountDialog from "../welcome/CreateAccountDialog";
+
+
 
 
                                                                 /* ==================== CONSTANTS ===================== */
@@ -112,60 +112,4 @@ const LoginForm = () => {
 	);
 };
 
-                                                                /* -------------------- LOGIN MENU -------------------- */
-const LoginMenu = () => {
-                                                                /* ----------------- State Variables ------------------ */
-    const [isSignupOpen, setIsSignupOpen] = useState(false);
-
-                                                                /* ------------------ Event Handlers ------------------ */
-                                                                /* Event handler for signup button                      */
-    const handleSignupOpen = () => {
-        setIsSignupOpen(true);                                  /* Show signup dialog                                   */          
-    };
-
-                                                                /* Event handler for signup dialog close                */
-    const handleSignupClose = () => {
-        setIsSignupOpen(false);                                 /* Hide signup dialog                                   */
-    };
-
-
-    return (
-        <Box
-            sx={{
-                padding: "20px"
-            }}
-        >
-            <LoginForm />
-            <Divider sx={{ margin: "20px" }} />
-            <Typography
-                variant="body1"
-                align="center"
-                sx={{ padding: "10px" }}
-            >
-                Don't have an account?
-            </Typography>
-            <Box
-                sx={{
-                    display: "flex",
-                    justifyContent: "center"
-                }}
-            >
-                <Button
-                    variant="contained"
-                    size="large"
-                    color="primary"
-                    sx={{ margin: "10px" }}
-                    onClick={handleSignupOpen}
-                >
-                    <Typography variant="body1">Sign Up</Typography>
-                </Button>
-            </Box>
-            <CreateAccountDialog onOpen={isSignupOpen} onClose={handleSignupClose} />
-        </Box>
-    );
-};
-
-
-                                                                /* ===================== EXPORTS ====================== */
-                                                                
-export default LoginMenu;
+export default LoginForm;
