+++ client/src/scenes/form/index.js
@@ -2,5 +2,6 @@ import LoginForm from "./LoginForm";
 import SignupForm from "./SignupForm";
 import PreparePortfolioForm from "./PreparePortfolioForm";
 import WatchSomeStocksForm from "./WatchSomeStocksForm";
+import LoginMenu from "./LoginMenu";
 
-export { LoginForm, SignupForm, PreparePortfolioForm, WatchSomeStocksForm };
+export { LoginForm, SignupForm, PreparePortfolioForm, LoginMenu, WatchSomeStocksForm };
